requirements.txt
playwright==1.49.0
python-telegram-bot==21.10
ollama==0.4.7
